update api login for user and expert

+++ src/middleware/auth/Oauth2.ts
@@ -1,33 +1,81 @@
 import IAccessTokenModel from '@src/models/cpAccessToken/IAccessTokenModel';
 import IAdminEntities from '@src/models/cpAdmin/IAdminModel';
 import ClientEntities from '@src/models/cpClient/IClientModel';
+import IExpertModel from '@src/models/cpExpert/IExpertModel';
+import IUserModel from '@src/models/cpUser/IUserModel';
 import AccessTokenRepository from '@src/repository/AccessTokenRepository';
 import AdminRepository from '@src/repository/AdminRepository';
+import ExpertRepository from '@src/repository/ExpertRepository';
+import UserRepository from '@src/repository/UserRepository';
 import {contants, security} from '@src/utils';
 import {randomBytes} from 'crypto';
 import {createServer, exchange, ExchangeDoneFunction} from 'oauth2orize';
 import passport from 'passport';
 
 // initialization token
-const initToken = async (client: ClientEntities, adminModel: IAdminEntities, done: ExchangeDoneFunction) => {
+const initToken = async (client: ClientEntities, clientModel: any, type: string, done: ExchangeDoneFunction) => {
   try {
     const _accessTokenRepository = new AccessTokenRepository();
-    const result = await _accessTokenRepository.findOne({
-      client_id: client.client_id,
-      id_admin: adminModel._id.toString(),
-    } as IAccessTokenModel);
-    if (result) {
-      await _accessTokenRepository.delete(result._id);
+
+    if (type === contants.TYPE_OF_CLIENT.ADMIN) {
+      const result = await _accessTokenRepository.findOne({
+        client_id: client.client_id,
+        id_client: clientModel._id.toString(),
+      } as IAccessTokenModel);
+      if (result) {
+        await _accessTokenRepository.delete(result._id);
+      }
+      const tokenValue = randomBytes(128).toString('hex');
+      await _accessTokenRepository.create({
+        client_id: client.client_id,
+        id_client: clientModel._id.toString(),
+        type: contants.TYPE_OF_CLIENT.ADMIN,
+        token: tokenValue,
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      } as IAccessTokenModel);
+      done(null, tokenValue);
     }
-    const tokenValue = randomBytes(128).toString('hex');
-    await _accessTokenRepository.create({
-      client_id: client.client_id,
-      id_admin: adminModel._id.toString(),
-      token: tokenValue,
-      createdAt: new Date(),
-      updatedAt: new Date(),
-    } as IAccessTokenModel);
-    done(null, tokenValue);
+
+    if (type === contants.TYPE_OF_CLIENT.USER) {
+      const result = await _accessTokenRepository.findOne({
+        client_id: client.client_id,
+        id_client: clientModel._id.toString(),
+      } as IAccessTokenModel);
+      if (result) {
+        await _accessTokenRepository.delete(result._id);
+      }
+      const tokenValue = randomBytes(128).toString('hex');
+      await _accessTokenRepository.create({
+        client_id: client.client_id,
+        id_client: clientModel._id.toString(),
+        type: contants.TYPE_OF_CLIENT.USER,
+        token: tokenValue,
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      } as IAccessTokenModel);
+      done(null, tokenValue);
+    }
+
+    if (type === contants.TYPE_OF_CLIENT.EXPERT) {
+      const result = await _accessTokenRepository.findOne({
+        client_id: client.client_id,
+        id_client: clientModel._id.toString(),
+      } as IAccessTokenModel);
+      if (result) {
+        await _accessTokenRepository.delete(result._id);
+      }
+      const tokenValue = randomBytes(128).toString('hex');
+      await _accessTokenRepository.create({
+        client_id: client.client_id,
+        id_client: clientModel._id.toString(),
+        type: contants.TYPE_OF_CLIENT.EXPERT,
+        token: tokenValue,
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      } as IAccessTokenModel);
+      done(null, tokenValue);
+    }
   } catch (error) {
     done(error);
   }
@@ -42,20 +90,47 @@ server.exchange(
     {},
     async (client, username: string, password: string, _scope, body, issused: ExchangeDoneFunction) => {
       try {
-        console.log(body, 'body');
-        const _adminRepository = new AdminRepository();
-        const admin = await _adminRepository.findOne({username} as IAdminEntities);
-        if (!admin) return issused(new Error('UseUserxist'));
+        if (body.type === contants.TYPE_OF_CLIENT.ADMIN) {
+          const _adminRepository = new AdminRepository();
+          const admin = await _adminRepository.findOne({username} as IAdminEntities);
+          if (!admin) return issused(new Error('UseAdminxist'));
 
-        if (!security.checkPassword(password.toString(), admin.salt.toString(), admin.hashed_password.toString()))
-          return issused(new Error('Login Fail'));
-        else {
-          if (admin.status === contants.STATUS.ACTIVE) {
-            // await admin.clearFCMToken(body.fcm_token);
-            // await cUser.updateById(cUseUser fcm_token: body.fcm_token });
-            initToken(client, admin, issused);
-          } else if (admin.status === contants.STATUS.DELETE) return issused(new Error('UseUserEEN_DELETED'));
-          else return issused(new Error('UserT_ACTIVE'));
+          if (!security.checkPassword(password.toString(), admin.salt.toString(), admin.hashed_password.toString()))
+            return issused(new Error('Login Fail'));
+          else {
+            if (admin.status === contants.STATUS.ACTIVE) {
+              initToken(client, admin, body.type, issused);
+            } else if (admin.status === contants.STATUS.DELETE) return issused(new Error('UseAdminEEN_DELETED'));
+            else return issused(new Error('AdminT_ACTIVE'));
+          }
+        } else if (body.type === contants.TYPE_OF_CLIENT.USER) {
+          const _userRepository = new UserRepository();
+          const user = await _userRepository.findOne({username} as IUserModel);
+          if (!user) return issused(new Error('UseUserxist'));
+
+          if (!security.checkPassword(password.toString(), user.salt.toString(), user.hashed_password.toString()))
+            return issused(new Error('Login Fail'));
+          else {
+            if (user.status === contants.STATUS.ACTIVE) {
+              initToken(client, user, body.type, issused);
+            } else if (user.status === contants.STATUS.DELETE) return issused(new Error('UseUserEEN_DELETED'));
+            else return issused(new Error('UserT_ACTIVE'));
+          }
+        } else if (body.type === contants.TYPE_OF_CLIENT.EXPERT) {
+          const _expertRepository = new ExpertRepository();
+          const expert = await _expertRepository.findOne({username} as IExpertModel);
+          if (!expert) return issused(new Error('UseExpertexist'));
+
+          if (!security.checkPassword(password.toString(), expert.salt.toString(), expert.hashed_password.toString()))
+            return issused(new Error('Login Fail'));
+          else {
+            if (expert.status === contants.STATUS.ACTIVE) {
+              initToken(client, expert, body.type, issused);
+            } else if (expert.status === contants.STATUS.DELETE) return issused(new Error('UseExpertEEN_DELETED'));
+            else return issused(new Error('ExpertT_ACTIVE'));
+          }
+        } else {
+          return issused(new Error('ACCOUNT_NOT_EXIST'));
         }
       } catch (error) {
         issused(error);
@@ -78,6 +153,7 @@ server.exchange(
  *
  * @apiParam {String} username
  * @apiParam {String} password
+ * @apiParam {String} type
  * @apiParam {String} grant_type password
  * @apiParam {String} client_id b109f3bbbc244eb82441917ed06d618b9008dd09b3bef
  * @apiParam {String} client_secret password
